+++ src/main/java/mctmods/immersivetechnology/api/ITUtils.java
@@ -2,7 +2,7 @@ package mctmods.immersivetechnology.api;
 
 import mctmods.immersivetechnology.api.client.MechanicalEnergyAnimation;
 import mctmods.immersivetechnology.common.blocks.ITBlockInterfaces.IMechanicalEnergy;
-import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntityAlternator;
+import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntityAlternatorMaster;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
@@ -11,78 +11,117 @@ import net.minecraft.world.World;
 import net.minecraftforge.fluids.IFluidTank;
 
 import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Set;
 
 public class ITUtils {
 	public static IFluidTank[] emptyIFluidTankList = new IFluidTank[0];
+
+	public static final Set<TileEntity> REMOVE_FROM_TICKING = new HashSet<>();
+
+	public static void RemoveDummyFromTicking(TileEntity te) {
+		REMOVE_FROM_TICKING.add(te);
+	}
 
 	public static BlockPos LocalOffsetToWorldBlockPos(BlockPos origin, int x, int y, int z, EnumFacing facing) {
 		return LocalOffsetToWorldBlockPos(origin, x, y, z, facing, EnumFacing.UP);
 	}
 
 	public static BlockPos LocalOffsetToWorldBlockPos(BlockPos origin, int x, int y, int z, EnumFacing facing, EnumFacing up) {
-		if(facing.getAxis() == up.getAxis()) throw new IllegalArgumentException("'facing' and 'up' must be perpendicular to each other!");
+		if (facing.getAxis() == up.getAxis())
+			throw new IllegalArgumentException("'facing' and 'up' must be perpendicular to each other!");
 		switch (up) {
 			case UP:
 				switch (facing) {
-					case SOUTH: return origin.add(-x, y, z);
-					case NORTH: return origin.add(x, y, -z);
-					case EAST: return origin.add(z, y, x);
-					case WEST: return origin.add(-z, y, -x);
-				default:
-					break;
-				} break;
+					case SOUTH:
+						return origin.add(-x, y, z);
+					case NORTH:
+						return origin.add(x, y, -z);
+					case EAST:
+						return origin.add(z, y, x);
+					case WEST:
+						return origin.add(-z, y, -x);
+					default:
+						break;
+				}
+				break;
 			case DOWN:
 				switch (facing) {
-					case SOUTH: return origin.add(x, -y, z);
-					case NORTH: return origin.add(-x, -y, -z);
-					case EAST: return origin.add(z, -y, -x);
-					case WEST: return origin.add(-z, -y, x);
-				default:
-					break;
-				} break;
+					case SOUTH:
+						return origin.add(x, -y, z);
+					case NORTH:
+						return origin.add(-x, -y, -z);
+					case EAST:
+						return origin.add(z, -y, -x);
+					case WEST:
+						return origin.add(-z, -y, x);
+					default:
+						break;
+				}
+				break;
 			case NORTH:
 				switch (facing) {
-					case UP: return origin.add(-x, z, -y);
-					case DOWN: return origin.add(x, -z, -y);
-					case EAST: return origin.add(z, x, -y);
-					case WEST: return origin.add(-z, -x, -y);
-				default:
-					break;
-				} break;
+					case UP:
+						return origin.add(-x, z, -y);
+					case DOWN:
+						return origin.add(x, -z, -y);
+					case EAST:
+						return origin.add(z, x, -y);
+					case WEST:
+						return origin.add(-z, -x, -y);
+					default:
+						break;
+				}
+				break;
 			case SOUTH:
 				switch (facing) {
-					case UP: return origin.add(x, z, y);
-					case DOWN: return origin.add(-x, -z, y);
-					case EAST: return origin.add(z, -x, y);
-					case WEST: return origin.add(-z, x, y);
-				default:
-					break;
-				} break;
+					case UP:
+						return origin.add(x, z, y);
+					case DOWN:
+						return origin.add(-x, -z, y);
+					case EAST:
+						return origin.add(z, -x, y);
+					case WEST:
+						return origin.add(-z, x, y);
+					default:
+						break;
+				}
+				break;
 			case EAST:
 				switch (facing) {
-					case UP: return origin.add(y, z, -x);
-					case DOWN: return origin.add(y, -z, x);
-					case SOUTH: return origin.add(y, x, z);
-					case NORTH: return origin.add(y, -x, -z);
-				default:
-					break;
-				} break;
+					case UP:
+						return origin.add(y, z, -x);
+					case DOWN:
+						return origin.add(y, -z, x);
+					case SOUTH:
+						return origin.add(y, x, z);
+					case NORTH:
+						return origin.add(y, -x, -z);
+					default:
+						break;
+				}
+				break;
 			case WEST:
 				switch (facing) {
-					case UP: return origin.add(-y, z, x);
-					case DOWN: return origin.add(-y, -z, -x);
-					case SOUTH: return origin.add(-y, -x, z);
-					case NORTH: return origin.add(-y, x, -z);
-				default:
-					break;
-				} break;
+					case UP:
+						return origin.add(-y, z, x);
+					case DOWN:
+						return origin.add(-y, -z, -x);
+					case SOUTH:
+						return origin.add(-y, -x, z);
+					case NORTH:
+						return origin.add(-y, x, -z);
+					default:
+						break;
+				}
+				break;
 		}
 		throw new IllegalArgumentException("This part of the code should never be reached! Has EnumFacing changed ? ");
 	}
 
-	public static <T> T First(ArrayList <T> list, Object o) {
-		for(T item: list) {
-			if(item.equals(o)) return item;
+	public static <T> T First(ArrayList<T> list, Object o) {
+		for (T item : list) {
+			if (item.equals(o)) return item;
 		}
 		return null;
 	}
@@ -93,22 +132,22 @@ public class ITUtils {
 		boundingArray[0] = fl == EnumFacing.WEST ? A : fl == EnumFacing.EAST ? B : fw == EnumFacing.EAST ? C : D;
 		boundingArray[1] = minY;
 		boundingArray[2] = fl == EnumFacing.NORTH ? A : fl == EnumFacing.SOUTH ? B : fw == EnumFacing.SOUTH ? C : D;
-		boundingArray[3] = fl == EnumFacing.EAST ? 1-A : fl == EnumFacing.WEST ? 1-B : fw == EnumFacing.EAST ? 1-D : 1-C;
+		boundingArray[3] = fl == EnumFacing.EAST ? 1 - A : fl == EnumFacing.WEST ? 1 - B : fw == EnumFacing.EAST ? 1 - D : 1 - C;
 		boundingArray[4] = maxY;
-		boundingArray[5] = fl == EnumFacing.SOUTH ? 1-A : fl == EnumFacing.NORTH ? 1-B : fw == EnumFacing.SOUTH ? 1-D : 1-C;
+		boundingArray[5] = fl == EnumFacing.SOUTH ? 1 - A : fl == EnumFacing.NORTH ? 1 - B : fw == EnumFacing.SOUTH ? 1 - D : 1 - C;
 
 		return boundingArray;
 	}
 
 	public static boolean checkMechanicalEnergyTransmitter(World world, BlockPos startPos) {
 		TileEntity tile = world.getTileEntity(startPos);
-		if(tile instanceof IMechanicalEnergy) {
-			if(((IMechanicalEnergy) tile).isMechanicalEnergyReceiver()) {
+		if (tile instanceof IMechanicalEnergy) {
+			if (((IMechanicalEnergy) tile).isMechanicalEnergyReceiver()) {
 				EnumFacing inputFacing = ((IMechanicalEnergy) tile).getMechanicalEnergyInputFacing();
 				BlockPos pos = startPos.offset(inputFacing, ((IMechanicalEnergy) tile).inputToCenterDistance() + 1);
 				TileEntity tileTransmitter = world.getTileEntity(pos);
 
-				if(tileTransmitter instanceof IMechanicalEnergy && ((IMechanicalEnergy) tileTransmitter).isMechanicalEnergyTransmitter() && (((IMechanicalEnergy) tileTransmitter).getMechanicalEnergyOutputFacing() == inputFacing.getOpposite())) {
+				if (tileTransmitter instanceof IMechanicalEnergy && ((IMechanicalEnergy) tileTransmitter).isMechanicalEnergyTransmitter() && (((IMechanicalEnergy) tileTransmitter).getMechanicalEnergyOutputFacing() == inputFacing.getOpposite())) {
 					return true;
 				}
 			}
@@ -119,13 +158,13 @@ public class ITUtils {
 	public static boolean checkMechanicalEnergyReceiver(World world, BlockPos startPos) {
 		TileEntity tile = world.getTileEntity(startPos);
 
-		if(tile instanceof IMechanicalEnergy) {
-			if(((IMechanicalEnergy) tile).isMechanicalEnergyTransmitter()) {
+		if (tile instanceof IMechanicalEnergy) {
+			if (((IMechanicalEnergy) tile).isMechanicalEnergyTransmitter()) {
 				EnumFacing outputFacing = ((IMechanicalEnergy) tile).getMechanicalEnergyOutputFacing();
 				BlockPos pos = startPos.offset(outputFacing, ((IMechanicalEnergy) tile).outputToCenterDistance() + 1);
 				TileEntity tileReceiver = world.getTileEntity(pos);
 
-				if(tileReceiver instanceof IMechanicalEnergy && ((IMechanicalEnergy) tileReceiver).isMechanicalEnergyReceiver() && ((IMechanicalEnergy) tileReceiver).getMechanicalEnergyInputFacing() == outputFacing.getOpposite()) {
+				if (tileReceiver instanceof IMechanicalEnergy && ((IMechanicalEnergy) tileReceiver).isMechanicalEnergyReceiver() && ((IMechanicalEnergy) tileReceiver).getMechanicalEnergyInputFacing() == outputFacing.getOpposite()) {
 					return true;
 				}
 			}
@@ -140,7 +179,7 @@ public class ITUtils {
 		TileEntity tileInfo = world.getTileEntity(pos);
 		TileEntity tileTransmitter = world.getTileEntity(pos.offset(inputFacing, ((IMechanicalEnergy) tileInfo).outputToCenterDistance()));
 
-		if(tileTransmitter instanceof IMechanicalEnergy) {
+		if (tileTransmitter instanceof IMechanicalEnergy) {
 			return ((IMechanicalEnergy) tileTransmitter).getEnergy();
 		} else {
 			return 0;
@@ -154,8 +193,8 @@ public class ITUtils {
 		TileEntity tileInfo = world.getTileEntity(pos);
 		TileEntity tileReceiver = world.getTileEntity(pos.offset(outputFacing, ((IMechanicalEnergy) tileInfo).inputToCenterDistance()));
 
-		if(tileReceiver instanceof TileEntityAlternator) {
-			if(((TileEntityAlternator) tileReceiver).canRunMechanicalEnergy()) {
+		if (tileReceiver instanceof TileEntityAlternatorMaster) {
+			if (((TileEntityAlternatorMaster) tileReceiver).canRunMechanicalEnergy()) {
 				return true;
 			}
 		}
@@ -177,7 +216,7 @@ public class ITUtils {
 		TileEntity tileInfo = world.getTileEntity(pos);
 		TileEntity tileTransmitter = world.getTileEntity(pos.offset(inputFacing, ((IMechanicalEnergy) tileInfo).outputToCenterDistance()));
 
-		if(tileTransmitter instanceof IMechanicalEnergy) {
+		if (tileTransmitter instanceof IMechanicalEnergy) {
 			return ((IMechanicalEnergy) tileTransmitter).getAnimation();
 		} else {
 			return new MechanicalEnergyAnimation();
@@ -188,17 +227,16 @@ public class ITUtils {
 		TileEntity tileTransmitter;
 		BlockPos pos;
 
-		for(EnumFacing f: EnumFacing.HORIZONTALS) {
+		for (EnumFacing f : EnumFacing.HORIZONTALS) {
 			pos = startPos.offset(f, 1);
 			tileTransmitter = world.getTileEntity(pos);
 
-			if(tileTransmitter instanceof IMechanicalEnergy) {
-				if(((IMechanicalEnergy) tileTransmitter).isMechanicalEnergyTransmitter() && ((IMechanicalEnergy) tileTransmitter).getMechanicalEnergyOutputFacing() == f.getOpposite()) {
+			if (tileTransmitter instanceof IMechanicalEnergy) {
+				if (((IMechanicalEnergy) tileTransmitter).isMechanicalEnergyTransmitter() && ((IMechanicalEnergy) tileTransmitter).getMechanicalEnergyOutputFacing() == f.getOpposite()) {
 					return f;
 				}
 			}
 		}
 		return null;
 	}
-
 }
