+++ src/main/java/mctmods/immersivetechnology/common/CommonProxy.java
@@ -3,6 +3,11 @@ package mctmods.immersivetechnology.common;
 import javax.annotation.Nonnull;
 
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.IGuiTile;
+import mctmods.immersivetechnology.api.ITUtils;
+import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntityBoilerMaster;
+import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntityDistillerMaster;
+import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntitySolarTowerMaster;
+import mctmods.immersivetechnology.common.blocks.stone.tileentities.TileEntityCokeOvenAdvancedMaster;
 import mctmods.immersivetechnology.ImmersiveTech;
 import mctmods.immersivetechnology.api.ITLib;
 import mctmods.immersivetechnology.client.gui.GuiBoiler;
@@ -12,11 +17,7 @@ import mctmods.immersivetechnology.client.gui.GuiSolarTower;
 import mctmods.immersivetechnology.client.gui.GuiTimer;
 import mctmods.immersivetechnology.client.gui.GuiTrashItem;
 import mctmods.immersivetechnology.common.blocks.connectors.tileentities.TileEntityTimer;
-import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntityBoiler;
-import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntityDistiller;
-import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntitySolarTower;
 import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntityTrashItem;
-import mctmods.immersivetechnology.common.blocks.stone.tileentities.TileEntityCokeOvenAdvanced;
 import mctmods.immersivetechnology.common.gui.ContainerBoiler;
 import mctmods.immersivetechnology.common.gui.ContainerCokeOvenAdvanced;
 import mctmods.immersivetechnology.common.gui.ContainerDistiller;
@@ -31,12 +32,25 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+import net.minecraftforge.fml.common.gameevent.TickEvent;
 import net.minecraftforge.fml.common.network.IGuiHandler;
 import net.minecraftforge.fml.relauncher.Side;
 
 public class CommonProxy implements IGuiHandler {
 
-	public void preInit() {}
+	public void preInit() {
+		MinecraftForge.EVENT_BUS.register(this);
+	}
+
+	@SubscribeEvent
+	public void onWorldTick(TickEvent.WorldTickEvent event) {
+		if (!ITUtils.REMOVE_FROM_TICKING.isEmpty() && event.phase == TickEvent.Phase.END) {
+			event.world.tickableTileEntities.removeAll(ITUtils.REMOVE_FROM_TICKING);
+			ITUtils.REMOVE_FROM_TICKING.clear();
+		}
+	}
 
 	public void preInitEnd() {}
 
@@ -61,10 +75,10 @@ public class CommonProxy implements IGuiHandler {
 		TileEntity tile = world.getTileEntity(new BlockPos(x, y, z));
 		if(tile instanceof IGuiTile) {
 			Object gui = null;
-			if(ID == ITLib.GUIID_Boiler && tile instanceof TileEntityBoiler) gui = new ContainerBoiler(player.inventory, (TileEntityBoiler) tile);
-			if(ID == ITLib.GUIID_Coke_oven_advanced && tile instanceof TileEntityCokeOvenAdvanced) gui = new ContainerCokeOvenAdvanced(player.inventory, (TileEntityCokeOvenAdvanced) tile);
-			if(ID == ITLib.GUIID_Distiller && tile instanceof TileEntityDistiller) gui = new ContainerDistiller(player.inventory, (TileEntityDistiller) tile);
-			if(ID == ITLib.GUIID_Solar_Tower && tile instanceof TileEntitySolarTower) gui = new ContainerSolarTower(player.inventory, (TileEntitySolarTower) tile);
+			if(ID == ITLib.GUIID_Boiler && tile instanceof TileEntityBoilerMaster) gui = new ContainerBoiler(player.inventory, (TileEntityBoilerMaster) tile);
+			if(ID == ITLib.GUIID_Coke_oven_advanced && tile instanceof TileEntityCokeOvenAdvancedMaster) gui = new ContainerCokeOvenAdvanced(player.inventory, (TileEntityCokeOvenAdvancedMaster) tile);
+			if(ID == ITLib.GUIID_Distiller && tile instanceof TileEntityDistillerMaster) gui = new ContainerDistiller(player.inventory, (TileEntityDistillerMaster) tile);
+			if(ID == ITLib.GUIID_Solar_Tower && tile instanceof TileEntitySolarTowerMaster) gui = new ContainerSolarTower(player.inventory, (TileEntitySolarTowerMaster) tile);
 			if(ID == ITLib.GUIID_Timer && tile instanceof TileEntityTimer) gui = new ContainerTimer(player.inventory, (TileEntityTimer) tile);
 			if(ID == ITLib.GUIID_Trash_Item && tile instanceof TileEntityTrashItem) gui = new ContainerTrashItem(player.inventory, (TileEntityTrashItem) tile);
 			if(gui != null) ((IGuiTile)tile).onGuiOpened(player, false);
@@ -78,10 +92,10 @@ public class CommonProxy implements IGuiHandler {
 		TileEntity tile = world.getTileEntity(new BlockPos(x, y, z));
 		if(tile instanceof IGuiTile) {
 			Object gui = null;
-			if(ID == ITLib.GUIID_Boiler && tile instanceof TileEntityBoiler) gui = new GuiBoiler(player.inventory, (TileEntityBoiler) tile);
-			if(ID == ITLib.GUIID_Coke_oven_advanced && tile instanceof TileEntityCokeOvenAdvanced) gui = new GuiCokeOvenAdvanced(player.inventory, (TileEntityCokeOvenAdvanced) tile);
-			if(ID == ITLib.GUIID_Distiller && tile instanceof TileEntityDistiller) gui = new GuiDistiller(player.inventory, (TileEntityDistiller) tile);
-			if(ID == ITLib.GUIID_Solar_Tower && tile instanceof TileEntitySolarTower) gui = new GuiSolarTower(player.inventory, (TileEntitySolarTower) tile);
+			if(ID == ITLib.GUIID_Boiler && tile instanceof TileEntityBoilerMaster) gui = new GuiBoiler(player.inventory, (TileEntityBoilerMaster) tile);
+			if(ID == ITLib.GUIID_Coke_oven_advanced && tile instanceof TileEntityCokeOvenAdvancedMaster) gui = new GuiCokeOvenAdvanced(player.inventory, (TileEntityCokeOvenAdvancedMaster) tile);
+			if(ID == ITLib.GUIID_Distiller && tile instanceof TileEntityDistillerMaster) gui = new GuiDistiller(player.inventory, (TileEntityDistillerMaster) tile);
+			if(ID == ITLib.GUIID_Solar_Tower && tile instanceof TileEntitySolarTowerMaster) gui = new GuiSolarTower(player.inventory, (TileEntitySolarTowerMaster) tile);
 			if(ID == ITLib.GUIID_Timer && tile instanceof TileEntityTimer) gui = new GuiTimer(player.inventory, (TileEntityTimer) tile);
 			if(ID == ITLib.GUIID_Trash_Item && tile instanceof TileEntityTrashItem) gui = new GuiTrashItem(player.inventory, (TileEntityTrashItem) tile);
 			return gui;
@@ -96,5 +110,4 @@ public class CommonProxy implements IGuiHandler {
 	public World getClientWorld() {
 		return null;
 	}
-
 }
