+++ src/main/java/mctmods/immersivetechnology/common/blocks/metal/multiblocks/MultiblockSolarTower.java
@@ -10,7 +10,7 @@ import blusunrize.immersiveengineering.common.blocks.metal.BlockTypes_MetalDecor
 import blusunrize.immersiveengineering.common.blocks.metal.BlockTypes_MetalDevice1;
 import blusunrize.immersiveengineering.common.util.Utils;
 import mctmods.immersivetechnology.common.ITContent;
-import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntitySolarTower;
+import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntitySolarTowerSlave;
 import mctmods.immersivetechnology.common.blocks.metal.types.BlockType_MetalMultiblock;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.GlStateManager;
@@ -84,10 +84,9 @@ public class MultiblockSolarTower implements IMultiblock {
 
 	@Override
 	public boolean createStructure(World world, BlockPos pos, EnumFacing side, EntityPlayer player) {
-		side = side.getOpposite();
-		if(side == EnumFacing.UP || side == EnumFacing.DOWN) {
-			side = EnumFacing.fromAngle(player.rotationYaw);
-		}
+		side = (side == EnumFacing.UP || side == EnumFacing.DOWN)? EnumFacing.fromAngle(player.rotationYaw) : side.getOpposite();
+		IBlockState master = ITContent.blockMetalMultiblock.getStateFromMeta(BlockType_MetalMultiblock.SOLAR_TOWER.getMeta());
+		IBlockState slave = ITContent.blockMetalMultiblock.getStateFromMeta(BlockType_MetalMultiblock.SOLAR_TOWER_SLAVE.getMeta());
 		boolean bool = this.structureCheck(world, pos, side);
 		if(!bool) return false;
 		for(int h = - 1 ; h <= 5 ; h ++) {
@@ -95,14 +94,15 @@ public class MultiblockSolarTower implements IMultiblock {
 				for(int w = - 1 ; w <= 1 ; w ++) {
 					if(h == 2 && ((l == 1 && (w == - 1 || w == 1)) || (w == 0 && (l == 0 || l == 2)))) continue;
 					BlockPos pos2 = pos.offset(side, l).offset(side.rotateY(), w).add(0, h, 0);
-					world.setBlockState(pos2, ITContent.blockMetalMultiblock.getStateFromMeta(BlockType_MetalMultiblock.SOLAR_TOWER.getMeta()));
+					int[] offset = new int[] {(side == EnumFacing.WEST ? - l : side == EnumFacing.EAST ? l : side == EnumFacing.NORTH ? w : - w), h, (side == EnumFacing.NORTH ? - l : side == EnumFacing.SOUTH ? l : side == EnumFacing.EAST ? w : - w)};
+					world.setBlockState(pos2, (offset[0]==0&&offset[1]==0&&offset[2]==0)? master : slave);
 					TileEntity curr = world.getTileEntity(pos2);
-					if(curr instanceof TileEntitySolarTower) {
-						TileEntitySolarTower tile = (TileEntitySolarTower)curr;
+					if(curr instanceof TileEntitySolarTowerSlave) {
+						TileEntitySolarTowerSlave tile = (TileEntitySolarTowerSlave)curr;
 						tile.facing = side;
 						tile.formed = true;
 						tile.pos = (h + 1) * 9 + l * 3 + (w + 1);
-						tile.offset = new int[] {(side == EnumFacing.WEST ? - l : side == EnumFacing.EAST ? l : side == EnumFacing.NORTH ? w : - w), h, (side == EnumFacing.NORTH ? - l : side == EnumFacing.SOUTH ? l : side == EnumFacing.EAST ? w : - w)};
+						tile.offset = offset;
 						tile.markDirty();
 						world.addBlockEvent(pos2, ITContent.blockMetalMultiblock, 255, 0);
 					}
