+++ src/main/java/mctmods/immersivetechnology/common/blocks/metal/tileentities/TileEntityAlternatorSlave.java
@@ -13,172 +13,69 @@ import blusunrize.immersiveengineering.api.energy.immersiveflux.IFluxProvider;
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.IAdvancedCollisionBounds;
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.IAdvancedSelectionBounds;
 import blusunrize.immersiveengineering.common.blocks.TileEntityMultiblockPart;
-import blusunrize.immersiveengineering.common.util.EnergyHelper;
 import blusunrize.immersiveengineering.common.util.EnergyHelper.*;
 import blusunrize.immersiveengineering.common.util.Utils;
 import blusunrize.immersiveengineering.api.IEEnums.SideConfig;
-import mctmods.immersivetechnology.ImmersiveTech;
 import mctmods.immersivetechnology.api.ITUtils;
 import mctmods.immersivetechnology.api.client.MechanicalEnergyAnimation;
-import mctmods.immersivetechnology.common.Config.ITConfig.MechanicalEnergy;
-import mctmods.immersivetechnology.common.Config.ITConfig.Machines.Alternator;
 import mctmods.immersivetechnology.common.blocks.ITBlockInterfaces.IMechanicalEnergy;
 import mctmods.immersivetechnology.common.blocks.metal.multiblocks.MultiblockAlternator;
-import mctmods.immersivetechnology.common.util.ITSounds;
-import mctmods.immersivetechnology.common.util.network.MessageStopSound;
-import mctmods.immersivetechnology.common.util.network.MessageTileSync;
-import mctmods.immersivetechnology.common.util.sound.ITSoundHandler;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.entity.EntityPlayerSP;
+import mctmods.immersivetechnology.common.blocks.stone.tileentities.TileEntityCokeOvenAdvancedMaster;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.SoundCategory;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
 
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.IFluidTank;
-import net.minecraftforge.fml.common.network.NetworkRegistry;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
 
-public class TileEntityAlternator extends TileEntityMultiblockPart <TileEntityAlternator> implements IMechanicalEnergy, IAdvancedSelectionBounds, IAdvancedCollisionBounds, IFluxProvider, IIEInternalFluxHandler {
+public class TileEntityAlternatorSlave extends TileEntityMultiblockPart <TileEntityAlternatorSlave> implements IMechanicalEnergy, IAdvancedSelectionBounds, IAdvancedCollisionBounds, IFluxProvider, IIEInternalFluxHandler {
 	private static int[] size = new int[] {3, 4, 3};
-	private static int maxSpeed = MechanicalEnergy.mechanicalEnergy_speed_max;
-	private static int rfPerTick = Alternator.alternator_energy_perTick;
-	private static int rfPerTickPerPort = rfPerTick / 6;
 
-	FluxStorage energyStorage = new FluxStorage(Alternator.alternator_energy_capacitorSize,rfPerTick,rfPerTickPerPort);
 
-	private BlockPos[] EnergyOutputPositions = new BlockPos[6];
-
-	public int speed;
-	private int clientUpdateCooldown = 20;
-	private float clientEnergyPercentage;
-	private int oldEnergy = energyStorage.getEnergyStored();
-	
-	MechanicalEnergyAnimation animation = new MechanicalEnergyAnimation();
-
-	public TileEntityAlternator() {
+	public TileEntityAlternatorSlave() {
 		super(size);
 	}
 
 	@Override
 	public void readCustomNBT(NBTTagCompound nbt, boolean descPacket) {
 		super.readCustomNBT(nbt, descPacket);
-		energyStorage.readFromNBT(nbt);
-		clientEnergyPercentage = (float) energyStorage.getEnergyStored() / energyStorage.getMaxEnergyStored();
-		speed = nbt.getInteger("speed");
-		animation.readFromNBT(nbt);
 	}
 
 	@Override
 	public void writeCustomNBT(NBTTagCompound nbt, boolean descPacket) {
 		super.writeCustomNBT(nbt, descPacket);
-		energyStorage.writeToNBT(nbt);
-		nbt.setInteger("speed", speed);
-		animation.writeToNBT(nbt);
-	}
-
-	public int energyGenerated() {
-		return Math.round(((float)speed / maxSpeed) * rfPerTick);
-	}
-
-	public void handleSounds() {
-		BlockPos center = getPos();
-		if(clientEnergyPercentage == 0) ITSoundHandler.StopSound(center);
-		else {
-			EntityPlayerSP player = Minecraft.getMinecraft().player;
-			float attenuation = Math.max((float) player.getDistanceSq(center.getX(), center.getY(), center.getZ()) / 8, 1);
-			ITSoundHandler.PlaySound(center, ITSounds.alternator, SoundCategory.BLOCKS, true, (2 * clientEnergyPercentage) / attenuation, clientEnergyPercentage);
-		}
-	}
-
-	@SideOnly(Side.CLIENT)
-	@Override
-	public void onChunkUnload() {
-		if(!isDummy()) ITSoundHandler.StopSound(getPos());
-		super.onChunkUnload();
-	}
-
-	@Override
-	public void disassemble() {
-		if(!isDummy()) {
-			BlockPos center = getPos();
-			ImmersiveTech.packetHandler.sendToAllTracking(new MessageStopSound(center), new NetworkRegistry.TargetPoint(world.provider.getDimension(), center.getX(), center.getY(), center.getZ(), 0));
-		}
-		super.disassemble();
-	}
-
-	public void notifyNearbyClients() {
-		NBTTagCompound tag = new NBTTagCompound();
-		tag.setInteger("energy", energyStorage.getEnergyStored());
-		BlockPos center = getPos();
-		ImmersiveTech.packetHandler.sendToAllTracking(new MessageTileSync(this, tag), new NetworkRegistry.TargetPoint(world.provider.getDimension(), center.getX(), center.getY(), center.getZ(), 0));
-	}
-
-	public void efficientMarkDirty() { // !!!!!!! only use it within update() function !!!!!!!
-		world.getChunkFromBlockCoords(this.getPos()).markDirty();
 	}
 
 	@Override
 	public void update() {
-		if(formed && pos == 13) {
-			if(!world.isRemote) {
-				if(ITUtils.checkMechanicalEnergyTransmitter(world, getPos())) speed = ITUtils.getMechanicalEnergy(world, getPos());
-				if(speed > 0) this.energyStorage.modifyEnergyStored(energyGenerated());
-				TileEntity tileEntity;
-				int currentEnergy = energyStorage.getEnergyStored();
-				for(int i = 0;i < 6;i++) {
-					if(currentEnergy == 0) break;
-					if(EnergyOutputPositions[i] == null) EnergyOutputPositions[i] = ITUtils.LocalOffsetToWorldBlockPos(getPos(), i < 3 ? -2 : 2, i < 3 ? i - 1 : i - 4, 0, facing);
-					tileEntity = Utils.getExistingTileEntity(world, EnergyOutputPositions[i]);
-					EnumFacing energyFacing = i < 3 ? facing.rotateY() : facing.rotateYCCW();
-					if(!EnergyHelper.isFluxReceiver(tileEntity, energyFacing)) continue;
-					int canReceiveAmount = EnergyHelper.insertFlux(tileEntity, energyFacing, Math.min(currentEnergy, rfPerTickPerPort), true);
-					if(canReceiveAmount == 0) continue;
-					EnergyHelper.insertFlux(tileEntity, energyFacing, canReceiveAmount, false);
-					energyStorage.modifyEnergyStored(-canReceiveAmount);
-					currentEnergy = energyStorage.getEnergyStored();
-				}
-				if (clientUpdateCooldown > 0) clientUpdateCooldown--;
-				if(oldEnergy != currentEnergy) {
-					efficientMarkDirty();
-					this.markContainingBlockForUpdate(null);
-					if (clientUpdateCooldown == 0) {
-						notifyNearbyClients();
-						clientUpdateCooldown = 20;
-					}
-				}
-				oldEnergy = currentEnergy;
-			} else handleSounds();
-		}
-	}
-
-	@Override
-	public void receiveMessageFromServer(NBTTagCompound message) {
-		if(message.hasKey("energy")) {
-			clientEnergyPercentage = (float) message.getInteger("energy") / energyStorage.getMaxEnergyStored();
-		}
+		ITUtils.RemoveDummyFromTicking(this);
 	}
 
 	@Override
 	public boolean isDummy() {
-		return pos != 13;
+		return true;
 	}
-	
-	IEForgeEnergyWrapper wrapper = new IEForgeEnergyWrapper(this, null);
 
-	private boolean isEnergyPos(@Nullable EnumFacing enumFacing) {
+	TileEntityAlternatorMaster master;
+
+	public TileEntityAlternatorMaster master() {
+		if (master != null && !master.tileEntityInvalid) return master;
+		BlockPos masterPos = getPos().add(-offset[0], -offset[1], -offset[2]);
+		TileEntity te = Utils.getExistingTileEntity(world, masterPos);
+		master = te instanceof TileEntityAlternatorMaster?(TileEntityAlternatorMaster)te: null;
+		return master;
+	}
+
+	public boolean isEnergyPos(@Nullable EnumFacing enumFacing) {
 		if(!this.formed || enumFacing == null || enumFacing == EnumFacing.DOWN || enumFacing == EnumFacing.UP) return false;
-		TileEntityAlternator master = this.master();
-		if(master == null) return false;
-		return (enumFacing.rotateY() == master().facing) && (pos == 0 || pos == 12 || pos == 24) ||
-				(enumFacing.rotateYCCW() == master().facing) && (pos == 2 || pos == 14 || pos == 26);
+		if(master() == null) return false;
+		return (enumFacing.rotateY() == master.facing) && (pos == 0 || pos == 12 || pos == 24) ||
+				(enumFacing.rotateYCCW() == master.facing) && (pos == 2 || pos == 14 || pos == 26);
 	}
 
 	@Nonnull
@@ -189,16 +86,15 @@ public class TileEntityAlternator extends TileEntityMultiblockPart <TileEntityAl
 
 	@Override
 	public IEForgeEnergyWrapper getCapabilityWrapper(EnumFacing facing) {
-		if(this.formed && this.isEnergyPos(facing)) return wrapper;
+		if(this.formed && this.isEnergyPos(facing) && master() != null) return master.wrapper;
 		return null;
 	}
 
 	@Nonnull
 	@Override
 	public FluxStorage getFluxStorage() {
-		TileEntityAlternator master = this.master();
-		if(master != null) return master.energyStorage;
-		return energyStorage;
+		if(master() != null) return master.energyStorage;
+		return new FluxStorage(0);
 	}
 
 	@Override
@@ -214,20 +110,17 @@ public class TileEntityAlternator extends TileEntityMultiblockPart <TileEntityAl
 	@Override
 	public int extractEnergy(@Nullable EnumFacing from, int energy, boolean simulate) {
 		if(!isEnergyPos(from)) return 0;
-		TileEntityAlternator master = master();
-		return master == null ? 0:master.energyStorage.extractEnergy(energy, simulate);
+		return master() == null ? 0 : master.energyStorage.extractEnergy(energy, simulate);
 	}
 
 	@Override
 	public int getEnergyStored(@Nullable EnumFacing from) {
-		TileEntityAlternator master = master();
-		return master == null ? 0:master.energyStorage.getEnergyStored();
+		return master() == null ? 0 : master.energyStorage.getEnergyStored();
 	}
 
 	@Override
 	public int getMaxEnergyStored(@Nullable EnumFacing from) {
-		TileEntityAlternator master = master();
-		return master == null ? 0:master.energyStorage.getMaxEnergyStored();
+		return master() == null ? 0 : master.energyStorage.getMaxEnergyStored();
 	}
 
 	public boolean canRunMechanicalEnergy() {
@@ -266,11 +159,11 @@ public class TileEntityAlternator extends TileEntityMultiblockPart <TileEntityAl
 
 	@Override
 	public int getEnergy() {
-		return speed;
+		return master() == null ? 0 : master.speed;
 	}
 
 	public MechanicalEnergyAnimation getAnimation() {
-		return animation;
+		return master() == null ? null : master.animation;
 	}
 
 	@Override
@@ -438,5 +331,4 @@ public class TileEntityAlternator extends TileEntityMultiblockPart <TileEntityAl
 		ArrayList <AxisAlignedBB> list) {
 		return false;
 	}
-
 }
