+++ src/main/java/mctmods/immersivetechnology/common/gui/ContainerSolarTower.java
@@ -2,7 +2,7 @@ package mctmods.immersivetechnology.common.gui;
 
 import blusunrize.immersiveengineering.common.gui.ContainerIEBase;
 import blusunrize.immersiveengineering.common.gui.IESlot;
-import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntitySolarTower;
+import mctmods.immersivetechnology.common.blocks.metal.tileentities.TileEntitySolarTowerMaster;
 import net.minecraft.entity.player.InventoryPlayer;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
@@ -11,11 +11,11 @@ import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 
-public class ContainerSolarTower extends ContainerIEBase<TileEntitySolarTower> {
-	public ContainerSolarTower(InventoryPlayer inventoryPlayer, TileEntitySolarTower tile) {
+public class ContainerSolarTower extends ContainerIEBase<TileEntitySolarTowerMaster> {
+	public ContainerSolarTower(InventoryPlayer inventoryPlayer, TileEntitySolarTowerMaster tile) {
 		super(inventoryPlayer, tile);
 
-		final TileEntitySolarTower tileF = tile;
+		final TileEntitySolarTowerMaster tileF = tile;
 		this.addSlotToContainer(new IESlot.FluidContainer(this, this.inv, 0, 80, 17, 2) {
 			@Override
 			public boolean isItemValid(ItemStack itemStack) {
