+++ src/main/resources/assets/immersivetech/blockstates/metal_multiblock.json
@@ -28,6 +28,24 @@
 				]
 			}
 		}],
+		"inventory,type=solar_tower_slave": [{
+			"model": "immersivetech:metal_multiblock/solar_tower.obj",
+			"transform": {
+				"scale": 0.125,
+				"rotation": [{
+					"x": 20
+				},
+					{
+						"y": -45
+					}
+				],
+				"translation": [
+					0,
+					-0.25,
+					-0.125
+				]
+			}
+		}],
 		"inventory,type=solar_reflector": [{
 			"model": "immersivetech:metal_multiblock/solar_reflector.obj",
 			"transform": {
@@ -35,6 +53,24 @@
 				"rotation": [{
 						"x": 20
 					},
+					{
+						"y": -45
+					}
+				],
+				"translation": [
+					0,
+					0,
+					0
+				]
+			}
+		}],
+		"inventory,type=solar_reflector_slave": [{
+			"model": "immersivetech:metal_multiblock/solar_reflector.obj",
+			"transform": {
+				"scale": 0.125,
+				"rotation": [{
+					"x": 20
+				},
 					{
 						"y": -45
 					}
@@ -64,15 +100,42 @@
 				]
 			}
 		}],
+		"inventory,type=alternator_slave": [{
+			"model": "immersivetech:metal_multiblock/alternator.obj",
+			"transform": {
+				"scale": 0.25,
+				"rotation": [{
+					"x": 20
+				},
+					{
+						"y": -45
+					}
+				],
+				"translation": [
+					0,
+					0,
+					0
+				]
+			}
+		}],
 		"type": {
 			"solar_tower": {
 				"model": "immersivetech:metal_multiblock/solar_tower.obj"
 			},
+			"solar_tower_slave": {
+				"model": "immersiveengineering:ie_empty"
+			},
 			"solar_reflector": {
 				"model": "immersivetech:metal_multiblock/solar_reflector.obj"
 			},
+			"solar_reflector_slave": {
+				"model": "immersiveengineering:ie_empty"
+			},
 			"alternator": {
 				"model": "immersivetech:metal_multiblock/alternator.obj"
+			},
+			"alternator_slave": {
+				"model": "immersiveengineering:ie_empty"
 			}
 		},
 		"facing": {
