+++ src/main/resources/assets/immersivetech/blockstates/metal_multiblock_distiller_slave.json
@@ -1,0 +1,80 @@
+{
+	"forge_marker": 1,
+	"defaults": {
+		"transform": "forge:default-block",
+		"custom": {
+			"flip-v": true
+		},
+		"textures": {
+			"particle": "immersiveengineering:blocks/storage_steel"
+		}
+	},
+	"variants": {
+		"inventory,type=distiller_slave": [{
+			"model": "immersivetech:metal_multiblock/distiller.obj",
+			"transform": {
+				"scale": 0.25,
+				"rotation": [{
+						"x": 20
+					},
+					{
+						"y": -45
+					}
+				]
+			}
+		}],
+		"type": {
+			"distiller_slave": {
+				"model": "immersiveengineering:ie_empty"
+			}
+		},
+		"facing": {
+			"north": {
+				"transform": {
+					"rotation": {
+						"y": 0
+					}
+				}
+			},
+			"south": {
+				"transform": {
+					"rotation": {
+						"y": 180
+					}
+				}
+			},
+			"west": {
+				"transform": {
+					"rotation": {
+						"y": 90
+					}
+				}
+			},
+			"east": {
+				"transform": {
+					"rotation": {
+						"y": -90
+					}
+				}
+			}
+		},
+		"_0multiblockslave": {
+			"false": {},
+			"true": {
+				"model": "immersiveengineering:ie_empty"
+			}
+		},
+		"_1dynamicrender": {
+			"false": {},
+			"true": {}
+		},
+		"boolean0": {
+			"false": {
+				"model": "immersivetech:metal_multiblock/distiller.obj"
+			},
+			"true": {
+				"model": "immersivetech:metal_multiblock/distiller_mirrored.obj"
+			}
+		}
+	}
+}
