+++ src/main/resources/assets/immersivetech/blockstates/metal_multiblock_steam_turbine_slave.json
@@ -1,0 +1,82 @@
+{
+  "forge_marker": 1,
+  "defaults": {
+    "transform": "forge:default-block",
+    "custom": {
+      "flip-v": true
+    },
+    "textures": {
+      "particle": "immersiveengineering:blocks/storage_steel"
+    }
+  },
+  "variants": {
+    "inventory,type=steam_turbine_slave": [{
+      "model": "immersivetech:metal_multiblock/steam_turbine.obj",
+      "transform": {
+        "scale": 0.25,
+        "rotation": [{
+          "x": 20
+        },
+          {
+            "y": -45
+          }
+        ]
+      }
+    }],
+    "type": {
+      "steam_turbine_slave": {
+        "model": "immersiveengineering:ie_empty"
+      }
+    },
+    "facing": {
+      "north": {
+        "transform": {
+          "rotation": {
+            "y": 0
+          }
+        }
+      },
+      "south": {
+        "transform": {
+          "rotation": {
+            "y": 180
+          }
+        }
+      },
+      "west": {
+        "transform": {
+          "rotation": {
+            "y": 90
+          }
+        }
+      },
+      "east": {
+        "transform": {
+          "rotation": {
+            "y": -90
+          }
+        }
+      }
+    },
+    "_0multiblockslave": {
+      "false": {},
+      "true": {
+        "model": "immersiveengineering:ie_empty"
+      }
+    },
+    "_1dynamicrender": {
+      "false": {},
+      "true": {
+        "model": "immersivetech:metal_multiblock/steam_turbine_rotor.obj"
+      }
+    },
+    "boolean0": {
+      "false": {
+        "model": "immersivetech:metal_multiblock/steam_turbine.obj"
+      },
+      "true": {
+        "model": "immersivetech:metal_multiblock/steam_turbine_mirrored.obj"
+      }
+    }
+  }
+}
