+++ src/main/resources/assets/immersivetech/blockstates/stone_multiblock.json
@@ -24,9 +24,26 @@
 				"translation": [0, 0, 0]
 			}
 		}],
+		"inventory,type=coke_oven_advanced_slave": [{
+			"model": "immersivetech:stone_multiblock/coke_oven_advanced.obj",
+			"transform": {
+				"scale": 0.25,
+				"rotation": [{
+					"x": 20
+				},
+					{
+						"y": -45
+					}
+				],
+				"translation": [0, 0, 0]
+			}
+		}],
 		"type": {
 			"coke_oven_advanced": {
 				"model": "immersivetech:stone_multiblock/coke_oven_advanced.obj"
+			},
+			"coke_oven_advanced_slave": {
+				"model": "immersiveengineering:ie_empty"
 			}
 		},
 		"facing": {
